+++ Rogue-Robots/DOGEngine/src/Audio/AudioPlayerComponent.h
@@ -2,7 +2,7 @@
 
 struct AudioPlayerComponent
 {
-	u64 audioID;
+	u32 audioID;
 	f32 volume = 2.0f;
 	// ...
 	i64 voiceID = -1;
